requirements.txt
python-telegram-bot[ext]
python-dotenv 